Add CASE_AADL_EventDataport connector type
This connector implements an EventDataport with 1 sender and N
receivers. Any instance can either be a native component or a VM
component. Each connection instance can be configured to generate
notification objects or instead leave the components to poll for
messages. Shared memory access permissions are read-write on the send
side and read-only on the receiver side.